Separate INBUS into auth, inbus, utils
