+++ common/inbus/inbus.py
@@ -1,87 +1,12 @@
-import requests
 import urllib.parse
 
-from oauthlib.oauth2 import BackendApplicationClient
-from requests_oauthlib import OAuth2Session
 from typing import Dict, Optional
 
-from django.conf import settings
-from django.core.cache import caches
-
-from . import dto
+from . import dto, utils
 
 INBUS_BASE_URL: str = 'https://inbus.vsb.cz/service/'
 INBUS_SERVICE_EDISON_URL: str = urllib.parse.urljoin(INBUS_BASE_URL, 'edison/v1/')
 INBUS_SERVICE_IDM_URL: str = urllib.parse.urljoin(INBUS_BASE_URL, 'idm/v1/')
-
-
-def authenticate(client_id=settings.INBUS_CLIENT_ID, client_secret=settings.INBUS_CLIENT_SECRET) -> Dict:
-    client = BackendApplicationClient(client_id=client_id)
-    oauth = OAuth2Session(client=client)
-    token = oauth.fetch_token(token_url='https://inbus.vsb.cz/oauth/token', client_id=client_id, client_secret=client_secret)
-
-    return token
-
-
-def set_token_to_cache(token: Dict) -> None:
-    """
-    Sets INBUS token to cache.
-    We set its timeout to one hour less than epecified by API provider.
-    """
-    cache = caches["default"]
-    timeout = token["expires_in"] - 3600 if token["expires_in"] > 3600 else token["expires_in"]
-    cache.set("inbus_token", token, timeout=timeout) # one hour less than provided
-
-
-def inbus_token() -> Dict:
-    """
-    Returns current INBUS token.
-    Either it's one that is cached or new one returned by authentication to INBUS.
-    """
-    cache = caches["default"]
-    token = cache.get("inbus_token")
-
-    if not token:
-        token = authenticate()
-        set_token_to_cache(token)
-    return token
-
-
-def request_new_token() -> Dict:
-    token = authenticate()
-    return token
-
-
-def is_response_ok_or_new_token_(response: requests.Response) -> bool:
-    if response.status_code == requests.codes.OK:
-        return True
-    elif response.status_code == requests.codes.UNAUTHORIZED:
-        token = request_new_token()
-        set_token_to_cache(token)
-        return False
-    else:
-        return False
-
-
-def inbus_request(url, params=None) -> Optional[requests.Response]:
-    if params is None:
-        params={}
-    token = inbus_token()
-    headers = {
-        "Authorization": f"Bearer {token['access_token']}",
-        "Accept" : "application/json",
-        "Accept-Language" : "cz"
-        }
-
-    response = requests.get(url, headers=headers, params=params)
-    if not is_response_ok_or_new_token_(response):
-        response = requests.get(url, headers=headers, params=params)
-        
-        # if we still don't get right response, fail with None
-        if response.status_code != requests.codes.OK:
-            return None
-
-    return response
 
 
 # Actual INBUS API calls
@@ -89,7 +14,7 @@ def inbus_request(url, params=None) -> Optional[requests.Response]:
 def person_by_login(login: str) -> Optional[dto.PersonSimple]:
     url = urllib.parse.urljoin(INBUS_SERVICE_IDM_URL, f'person/login/{login}')
 
-    person_resp = inbus_request(url, {})
+    person_resp = utils.inbus_request(url, {})
     if not person_resp:
         return None
     person_json = person_resp.json()
